+++ src/main/java/lotto/model/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import static lotto.model.constant.LottoPrizeConstant.FIFTH;
 import static lotto.model.constant.LottoPrizeConstant.FIRST;
